backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
python-dotenv>=1.0.1
pymongo==4.5.0
motor==3.3.1
pydantic>=2.6.4
python-multipart>=0.0.9
jq>=1.6.0
typer>=0.9.0
scikit-learn>=1.4.0
joblib>=1.3.0
Pillow>=10.0.0
pandas>=2.1.0
numpy>=1.24.0
nltk>=3.8.1
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
requests>=2.31.0
matplotlib>=3.7.0
seaborn>=0.12.0
# Nuevas librerías para ML avanzado
spacy>=3.7.0
textblob>=0.17.0
wordcloud>=1.9.0
plotly>=5.15.0